requirements.txt
torch
torchvision
pillow
numpy
scikit-learn
tqdm
open-clip-torch
